requirements.txt
torch>=2.5.1
wandb>=0.18.7
numpy==1.26.4
pandas>=2.2.3
tqdm>=4.67.0
hydra-core>=1.3.2
omegaconf==2.3.0